+++ decidim-core/app/helpers/decidim/check_boxes_tree_helper.rb
@@ -83,9 +83,13 @@ module Decidim
     end
 
     def filter_scopes_values
-      main_scopes = current_participatory_space.scope.present? ? [current_participatory_space.scope] : current_participatory_space.scopes.top_level
+      main_scopes = if current_component.scope
+                      [current_component.scope]
+                    else
+                      current_component.parent_component_or_space.scopes.includes(:scope_type, :children).top_level
+                    end
 
-      scopes_values = main_scopes.includes(:scope_type, :children).flat_map do |scope|
+      scopes_values = main_scopes.flat_map do |scope|
         TreeNode.new(
           TreePoint.new(scope.id.to_s, translated_attribute(scope.name, current_participatory_space.organization)),
           scope_children_to_tree(scope)