+++ decidim-meetings/app/forms/decidim/meetings/admin/meeting_form.rb
@@ -35,11 +35,12 @@ module Decidim
         validates :current_component, presence: true
         validates :category, presence: true, if: ->(form) { form.decidim_category_id.present? }
         validates :scope, presence: true, if: ->(form) { form.decidim_scope_id.present? }
+        validates :decidim_scope_id, subscope_belongs_to_component: true, if: ->(form) { form.decidim_scope_id.present? }
         validates :organizer, presence: true, if: ->(form) { form.organizer_id.present? }
 
-        validate :scope_belongs_to_participatory_space_scope
-
         delegate :categories, to: :current_component
+
+        alias component current_component
 
         def map_model(model)
           self.services = model.services.map do |service|
@@ -64,13 +65,11 @@ module Decidim
           @organizer ||= current_organization.users.find_by(id: organizer_id)
         end
 
-        alias component current_component
-
-        # Finds the Scope from the given decidim_scope_id, uses participatory space scope if missing.
+        # Finds the Scope from the given decidim_scope_id, uses component scope if missing.
         #
         # Returns a Decidim::Scope
         def scope
-          @scope ||= @decidim_scope_id ? current_participatory_space.scopes.find_by(id: @decidim_scope_id) : current_participatory_space.scope
+          @scope ||= @decidim_scope_id ? current_component.scopes.find_by(id: @decidim_scope_id) : current_component.scope
         end
 
         # Scope identifier
@@ -84,12 +83,6 @@ module Decidim
           return unless current_component
 
           @category ||= categories.find_by(id: decidim_category_id)
-        end
-
-        private
-
-        def scope_belongs_to_participatory_space_scope
-          errors.add(:decidim_scope_id, :invalid) if current_participatory_space.out_of_scope?(scope)
         end
       end
     end